Tornando adicionar-disciplina um método para todos os usuários

+++ back/src/users/users.service.ts
@@ -10,14 +10,12 @@ import { User } from './entities/user.entity';
 import { InjectModel } from '@nestjs/mongoose';
 import { Disciplina } from 'src/disciplinas/entities/disciplina.entity';
 import { hash, genSalt, compare } from 'bcrypt';
-import { AuthService } from 'src/auth/auth.service';
 
 @Injectable()
 export class UsersService {
   constructor(
     @InjectModel(User.name) private userModel: Model<User>,
     @InjectModel(Disciplina.name) private disciplinaModel: Model<Disciplina>,
-    private authService: AuthService,
   ) {}
 
   async create(createUserDto: CreateUserDto) {
@@ -58,8 +56,18 @@ export class UsersService {
     return this.userModel.updateOne({ _id: id }, updateUserDto).exec();
   }
 
-  remove(id: string) {
-    return this.userModel.deleteOne({ _id: id }).exec();
+  async remove(id: string) {
+    let usuario = await this.findOne(id);
+    if (usuario instanceof NotFoundException || usuario == null) return usuario;
+    if (usuario.disciplinas.length > 0) {
+      usuario.disciplinas.forEach(async (disciplina) => {
+        await this.removerDisciplina(id, disciplina.codigo_disciplina);
+      });
+    }
+
+    await this.userModel.deleteOne({ _id: id }).exec();
+
+    return { message: 'Usuário removido com sucesso!' };
   }
 
   async adicionarDisciplina(id_user: string, codigo_disciplina: string) {
@@ -73,24 +81,35 @@ export class UsersService {
 
     if (disciplina instanceof NotFoundException || disciplina == null)
       return disciplina;
-    if (usuario.e_professor == true)
-      throw new ConflictException(
-        'Um professor não pode cursar uma disciplina.',
-      );
 
     const { professores, alunos, atividades, ...filtro_disciplina } =
       disciplina.toJSON();
-    const { disciplinas, e_professor, senha, ...filtro_aluno } =
-      usuario.toJSON();
 
-    for (const aluno of disciplina.alunos) {
-      if (aluno._id.match(id_user)) {
-        throw new ConflictException('Aluno já cursa essa disciplina.');
+    if (usuario.e_professor == true) {
+      const { disciplinas, senha, matricula, ano, ...filtro_professor } =
+        usuario.toJSON();
+
+      for (const professor of disciplina.alunos) {
+        if (professor._id.match(id_user)) {
+          throw new ConflictException('Professor já leciona essa disciplina.');
+        }
       }
-    }
 
-    usuario.disciplinas.push(filtro_disciplina as any);
-    disciplina.alunos.push(filtro_aluno as any);
+      usuario.disciplinas.push(filtro_disciplina as any);
+      disciplina.professores.push(filtro_professor as any);
+    } else {
+      const { disciplinas, e_professor, senha, ...filtro_aluno } =
+        usuario.toJSON();
+
+      for (const aluno of disciplina.alunos) {
+        if (aluno._id.match(id_user)) {
+          throw new ConflictException('Aluno já cursa essa disciplina.');
+        }
+      }
+
+      usuario.disciplinas.push(filtro_disciplina as any);
+      disciplina.alunos.push(filtro_aluno as any);
+    }
 
     await Promise.all([disciplina.save(), usuario.save()]);
 
@@ -117,47 +136,62 @@ export class UsersService {
     if (disciplina instanceof NotFoundException || disciplina == null)
       return disciplina;
 
-    if (usuario.e_professor == true)
-      throw new ConflictException('Um professor não pode remover disciplinas.');
-
     const { professores, alunos, atividades, ...filtro_disciplina } =
       disciplina.toJSON();
     const { disciplinas, e_professor, senha, ...filtro_aluno } =
       usuario.toJSON();
 
-    const alunoIndex = disciplina.alunos.findIndex((aluno) =>
-      aluno._id.match(id_user),
-    );
-
-    if (alunoIndex === -1) {
-      throw new ConflictException(
-        'Aluno não está matriculado nessa disciplina.',
+    if (usuario.e_professor == true) {
+      const professor = disciplina.alunos.findIndex((aluno) =>
+        aluno._id.match(id_user),
       );
+
+      if (professor === -1) {
+        throw new ConflictException('Professor não leciona esta disciplina.');
+      }
+
+      disciplina.professores.splice(professor, 1);
+
+      const disciplinaIndex = usuario.disciplinas.findIndex(
+        (disc) => disc.codigo_disciplina === codigo_disciplina,
+      );
+
+      if (disciplinaIndex === -1) {
+        throw new ConflictException(
+          'Disciplina não está associada ao professor.',
+        );
+      }
+
+      usuario.disciplinas.splice(disciplinaIndex, 1);
+    } else {
+      const alunoIndex = disciplina.alunos.findIndex((aluno) =>
+        aluno._id.match(id_user),
+      );
+
+      if (alunoIndex === -1) {
+        throw new ConflictException(
+          'Aluno não está matriculado nessa disciplina.',
+        );
+      }
+
+      disciplina.alunos.splice(alunoIndex, 1);
+
+      const disciplinaIndex = usuario.disciplinas.findIndex(
+        (disc) => disc.codigo_disciplina === codigo_disciplina,
+      );
+
+      if (disciplinaIndex === -1) {
+        throw new ConflictException('Disciplina não está associada ao aluno.');
+      }
+
+      usuario.disciplinas.splice(disciplinaIndex, 1);
     }
-
-    disciplina.alunos.splice(alunoIndex, 1);
-
-    const disciplinaIndex = usuario.disciplinas.findIndex(
-      (disc) => disc.codigo_disciplina === codigo_disciplina,
-    );
-
-    if (disciplinaIndex === -1) {
-      throw new ConflictException('Disciplina não está associada ao aluno.');
-    }
-
-    usuario.disciplinas.splice(disciplinaIndex, 1);
 
     await Promise.all([disciplina.save(), usuario.save()]);
 
-    const {
-      email: retirar_email,
-      senha: retirar_senha,
-      ano: retirar_ano,
-      e_professor: retirar_e_professor,
-      ...retornar_usuario
-    } = usuario.toJSON();
-
-    return retornar_usuario;
+    return {
+      message: 'Usuário removido da disciplina com sucesso!',
+    };
   }
 
   async alterarSenha(email: string, senha_nova: string) {
@@ -169,10 +203,12 @@ export class UsersService {
     usuario_alterar.senha = await this.hashSenha(senha_nova);
 
     if (usuario_alterar?._id) {
-      await this.userModel.updateOne(
-        { _id: usuario_alterar?._id },
-        { senha: usuario_alterar.senha },
-      ).exec();
+      await this.userModel
+        .updateOne(
+          { _id: usuario_alterar?._id },
+          { senha: usuario_alterar.senha },
+        )
+        .exec();
       return { message: 'Senha alterada com sucesso!' };
     } else {
       return { message: 'Erro ao alterar senha!' };
